make generate_unique_output
